+++ include/asm-sparc/hdreg.h
@@ -1,4 +1,4 @@
-/* $Id: hdreg.h,v 1.1 2000/01/21 04:56:27 zaitcev Exp $
+/* $Id: hdreg.h,v 1.2 2000/12/05 00:56:36 anton Exp $
  * hdreg.h: SPARC PCI specific IDE glue.
  *
  * Copyright (C) 1997  David S. Miller (davem@caip.rutgers.edu)
@@ -8,6 +8,6 @@
 #ifndef __SPARC_HDREG_H
 #define __SPARC_HDREG_H
 
-typedef unsigned int ide_ioreg_t;
+typedef unsigned long ide_ioreg_t;
 
 #endif /* __SPARC_HDREG_H */
